Split wallet_committee.cpp into 2 files
