+++ libraries/wallet/CMakeLists.txt
@@ -37,9 +37,10 @@ set( SOURCES
      wallet_api_impl.cpp
      wallet_asset.cpp
      wallet_builder.cpp
-     wallet_committee.cpp
      wallet_debug.cpp
+     wallet_fee_change.cpp
      wallet_network.cpp
+     wallet_param_change.cpp
      wallet_results.cpp
      wallet_sign.cpp
      wallet_transfer.cpp
